docs(Phase 1): 1.5: Use openchami-certificate-update to configure certs
Instead of manually editing files to edit the FQDN used in certificate
generation and deployment, use the openchami-certificate-update script
included in OpenCHAMI v0.0.28.

It was easy to miss things when editing manually (one had to edit files
in /etc/openchami _and_ /etc/containers/systemd) which caused a lot of
issues. The script makes it quite a bit easier and more robust.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>